+++ androidx/src/androidTest/java/me/panpf/androidx/test/app/ActivityxTest.java
@@ -23,6 +23,7 @@ import android.os.Build;
 import android.os.Bundle;
 import android.view.View;
 
+import androidx.activity.ComponentActivity;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
@@ -181,12 +182,12 @@ public class ActivityxTest {
     @Test
     public void testGetImplWithParent() {
         TestActivity activity = activityTestRule.getActivity();
-        Assert.assertNotNull(Activityx.getImplWithParent(activity, ImplTestInterface.class));
+        Assert.assertEquals(Premisex.requireNotNull(Activityx.getImplWithParent(activity, ImplTestInterface.class)).getClass(), TestActivity.class);
         Assert.assertNull(Activityx.getImplWithParent(activity, ViewModelStoreOwner.class));
 
         TestFragmentActivity activity2 = fragmentActivityTestRule.getActivity();
         Assert.assertNull(Activityx.getImplWithParent(activity2, ImplTestInterface.class));
-        Assert.assertNotNull(Activityx.getImplWithParent(activity2, ViewModelStoreOwner.class));
+        Assert.assertEquals(Premisex.requireNotNull(Activityx.getImplWithParent(activity2, ViewModelStoreOwner.class)).getClass(), TestFragmentActivity.class);
     }
 
     @Test